Started making the website responsive using bootstrap. -Yamila
Reason being that the website looked very different on 2 laptops of different sizes, this could give problems because of the many differences in sizes in laptops and also the screensize of the bin pc's.
When completely finished, the website should adapt to different screensizes, leaving a clean look no matter what device the website is used on.

- Had to redo the basic frame of the website and rewrite it using a grid system, which should make the website responsive to different screen sizes (phone, laptop, tablet etc.)

- Removed all margins and most of the padding as they are not needed while using flex-grid system. The only remaining padding is used for hover nav_bar.

To Do: Still have to merge some of the css/bootstrap and decide exact measurements within grid system, to see which one works best.

(background color of area with boxes will change and was just used for me to easily be able to identify the region)